+++ src/app/login/login.component.ts
@@ -1,6 +1,8 @@
-import { Component, OnInit } from '@angular/core';
-import { Router } from '@angular/router';
-import { AuthenticationService } from '../../services/authentication.service';
+import {Component, OnInit} from '@angular/core';
+import {Router} from '@angular/router';
+import {AuthenticationService} from '../../services/authentication.service';
+import {FormControl, Validators, FormGroup} from '@angular/forms';
+import {LoginErrorStateMatcher} from './login_errors/login_error_state_matcher';
 
 @Component({
   moduleId: module.id,
@@ -9,15 +11,23 @@ import { AuthenticationService } from '../../services/authentication.service';
 })
 
 export class LoginComponent implements OnInit {
+  login_form = new FormGroup({
+    email: new FormControl('', [
+      Validators.pattern('^[a-z0-9._%+-]+@[a-z0-9.-]+\\.[a-z]{2,6}$'),
+      Validators.required
+    ]),
+    password: new FormControl('', [Validators.required, Validators.minLength(4)]),
+  });
+  matcher = new LoginErrorStateMatcher();
   model: any = {};
   loading = false;
   load_form = false;
+  hide = true;
   error = '';
 
-  constructor(
-    private router: Router,
-    private authenticationService: AuthenticationService,
-  ) { }
+  constructor(private router: Router,
+              private authenticationService: AuthenticationService,) {
+  }
 
   ngOnInit() {
     // reset login status
@@ -35,7 +45,7 @@ export class LoginComponent implements OnInit {
     this.loading = true;
     this.authenticationService.login(this.model.username, this.model.password)
       .then(result => {
-          this.router.navigate(['/']);
+        this.router.navigate(['/']);
         this.loading = false;
       }, error => {
         this.error = 'Username or password is incorrect';